Feat: update train.py files
